Refactor RealtimeService into modular subclasses by event domain #213
