+++ api/openai_realtime/handlers/base.py
@@ -1,0 +1,55 @@
+from __future__ import annotations
+
+import logging
+import uuid
+from queue import Queue
+from threading import Event as ThreadingEvent
+from typing import TYPE_CHECKING
+
+from openai.types.realtime import RealtimeErrorEvent
+
+from api.openai_realtime.runtime_config import RuntimeConfig
+
+if TYPE_CHECKING:
+    from api.openai_realtime.service import ConnState, RealtimeService
+
+logger = logging.getLogger(__name__)
+
+
+def _generate_id(prefix: str) -> str:
+    return f"{prefix}_{uuid.uuid4().hex[:12]}"
+
+
+class RealtimeBaseHandler:
+    """Shared base for domain handlers.
+
+    Provides conn_id-keyed accessors for per-connection state, config,
+    and queues.  Today these return the service-wide singletons; when
+    multi-session lands, only these accessors need to change.
+    """
+
+    def __init__(self, service: RealtimeService) -> None:
+        self._service = service
+
+    # ── conn_id-keyed accessors (multi-session seam) ──
+
+    def _config(self, conn_id: str) -> RuntimeConfig:
+        return self._service.runtime_config
+
+    def _queue(self, conn_id: str) -> Queue | None:
+        return self._service.text_prompt_queue
+
+    def _should_listen(self, conn_id: str) -> ThreadingEvent | None:
+        return self._service.should_listen
+
+    # ── shared helpers ────────────────────────────────
+
+    def _state(self, conn_id: str) -> ConnState:
+        return self._service._state(conn_id)
+
+    @staticmethod
+    def _next_event_id() -> str:
+        return _generate_id("event")
+
+    def make_error(self, message: str, _type: str) -> RealtimeErrorEvent:
+        return self._service.make_error(message, _type)
