+++ api/openai_realtime/service.py
@@ -1,23 +1,10 @@
-import base64
-import json
 import logging
-import uuid
 from pydantic import BaseModel, Field
 from queue import Queue
 from threading import Event as ThreadingEvent
-from typing import Literal, Optional, Union
+from typing import Callable, Literal, Optional, Union
 
 from pydantic import ValidationError
-
-from openai.types.realtime.realtime_response import Audio, AudioOutput
-from openai.types.realtime.realtime_response_status import RealtimeResponseStatus
-from openai.types.realtime.realtime_response_usage import RealtimeResponseUsage
-from openai.types.realtime.conversation_item_input_audio_transcription_completed_event import (
-    UsageTranscriptTextUsageDuration,
-)
-from openai.types.realtime.realtime_transcription_session_create_request import (
-    RealtimeTranscriptionSessionCreateRequest,
-)
 
 from openai.types.realtime import (
     InputAudioBufferAppendEvent,
@@ -27,7 +14,6 @@ from openai.types.realtime import (
     ResponseCreateEvent,
     ResponseCancelEvent,
     SessionCreatedEvent,
-    RealtimeSessionCreateRequest,
     RealtimeErrorEvent,
     RealtimeError,
     InputAudioBufferSpeechStartedEvent,
@@ -35,18 +21,21 @@ from openai.types.realtime import (
     ConversationItemInputAudioTranscriptionCompletedEvent,
     ConversationItemInputAudioTranscriptionDeltaEvent,
     ResponseCreatedEvent,
-    RealtimeResponse,
     ResponseDoneEvent,
     ResponseAudioDeltaEvent,
     ResponseAudioDoneEvent,
     ResponseAudioTranscriptDoneEvent,
     ResponseFunctionCallArgumentsDoneEvent,
-    ConversationItem
 )
 
 from api.openai_realtime.runtime_config import RuntimeConfig
-from api.openai_realtime.utils import resample
-from pipeline_messages import MessageTag
+from api.openai_realtime.handlers import (
+    AudioHandler,
+    ConversationHandler,
+    ResponseHandler,
+    SessionHandler,
+)
+from api.openai_realtime.handlers.base import _generate_id
 
 logger = logging.getLogger(__name__)
 
@@ -87,14 +76,10 @@ ServerEvent = Union[
     ResponseAudioDeltaEvent,
     ResponseAudioDoneEvent,
     ResponseAudioTranscriptDoneEvent,
-    ResponseFunctionCallArgumentsDoneEvent
+    ResponseFunctionCallArgumentsDoneEvent,
 ]
 
 RealtimeEvent = Union[ClientEvent, ServerEvent]
-
-def _generate_id(prefix: str) -> str:
-    return f"{prefix}_{uuid.uuid4().hex[:12]}"
-
 
 
 class UsageMetrics(BaseModel):
@@ -173,6 +158,20 @@ class RealtimeService:
         self._conns: dict[str, ConnState] = {}
         self.total_usage = GlobalUsageMetrics()
 
+        self.audio = AudioHandler(self)
+        self.session = SessionHandler(self)
+        self.response = ResponseHandler(self)
+        self.conversation = ConversationHandler(self)
+
+        self._pipeline_dispatch: dict[str, Callable[[str, dict], list[ServerEvent]]] = {
+            "speech_started": self.audio.on_speech_started,
+            "speech_stopped": self.audio.on_speech_stopped,
+            "assistant_text": self.response.on_assistant_text,
+            "token_usage": self._on_token_usage,
+            "partial_transcription": self.conversation.on_partial_transcription,
+            "transcription_completed": self.conversation.on_transcription_completed,
+        }
+
     # ── Connection lifecycle ─────────────────────
 
     def register(self) -> str:
@@ -199,97 +198,11 @@ class RealtimeService:
     def connection_ids(self) -> list[str]:
         return list(self._conns)
 
-    def build_session_created(self, conn_id: str) -> SessionCreatedEvent:
-        """Build a SessionCreatedEvent populated with the current RuntimeConfig."""
-        session = self.runtime_config.session or RealtimeSessionCreateRequest(type="realtime")
-        return SessionCreatedEvent(
-            type="session.created",
-            event_id=self._next_event_id(),
-            session=session,
-        )
-
-    # ── ID management ────────────────────────────
+    # ── Client event parsing ─────────────────────
 
     @staticmethod
     def _next_event_id() -> str:
         return _generate_id("event")
-
-    def _ensure_response(self, conn_id: str) -> tuple[str, str]:
-        """Ensure a response and output item exist, creating them if needed."""
-        st = self._state(conn_id)
-        if st.current_response_id is None:
-            st.current_response_id = _generate_id("resp")
-            self._start_item(conn_id)
-            st.in_response = True
-        return st.current_response_id, self._current_item_id(conn_id)
-
-    def _end_response(self, conn_id: str, status: _ResponseStatus = "completed") -> None:
-        st = self._state(conn_id)
-        if status == "cancelled":
-            st.response_usage.responses_cancelled += 1
-        else:
-            st.response_usage.responses_completed += 1
-        self.total_usage += st.response_usage
-        logger.info(
-            "Response done (status=%s) — this response: input_tokens=%d, output_tokens=%d, audio=%.2fs"
-            " | cumulative: input_tokens=%d, output_tokens=%d, audio=%.2fs",
-            status,
-            st.response_usage.input_tokens, st.response_usage.output_tokens,
-            st.response_usage.audio_duration_s,
-            self.total_usage.input_tokens, self.total_usage.output_tokens,
-            self.total_usage.audio_duration_s,
-        )
-        st.response_usage.reset()
-        st.current_response_id = None
-        st.current_item_id = None
-        st.content_index = 0
-        st.in_response = False
-        st.response_metadata = None
-
-    def _start_item(self, conn_id: str) -> str:
-        """Generate a new item ID, reset content index, and store it."""
-        st = self._state(conn_id)
-        item_id = _generate_id("item")
-        st.current_item_id = item_id
-        st.content_index = 0
-        st.input_audio_duration_s = 0.0
-        return item_id
-
-    def _current_item_id(self, conn_id: str) -> str:
-        return self._state(conn_id).current_item_id or self._start_item(conn_id)
-
-    def _next_content_index(self, conn_id: str) -> int:
-        """Return the current content index and advance it."""
-        st = self._state(conn_id)
-        idx = st.content_index
-        st.content_index += 1
-        return idx
-
-    def _build_response(self, conn_id: str, status: _ResponseStatus, reason: _StatusReason | None = None) -> RealtimeResponse:
-        """Build a fully-populated RealtimeResponse from the current connection state."""
-        st = self._state(conn_id)
-        status_details = None
-        if reason or status in ("completed", "cancelled", "incomplete", "failed"):
-            status_details = RealtimeResponseStatus(type=status, reason=reason)  # type: ignore[arg-type]
-        audio_cfg = self.runtime_config.session.audio
-        audio_output = audio_cfg.output if audio_cfg is not None else None
-        voice = audio_output.voice if audio_output is not None else None
-        return RealtimeResponse(
-            id=st.current_response_id,
-            object="realtime.response",
-            status=status,
-            status_details=status_details,
-            audio=Audio(output=AudioOutput(voice=voice)),
-            conversation_id=st.conversation_id,
-            metadata=st.response_metadata,
-            usage=RealtimeResponseUsage(
-                input_tokens=st.response_usage.input_tokens,
-                output_tokens=st.response_usage.output_tokens,
-                total_tokens=st.response_usage.input_tokens + st.response_usage.output_tokens,
-            ),
-        )
-
-    # ── Client event parsing ─────────────────────
 
     def parse_client_event(self, raw: dict) -> Optional[ClientEvent]:
         event_type: str | None = raw.get("type")
@@ -308,398 +221,56 @@ class RealtimeService:
 
     # ── Client event handlers ────────────────────
 
+    def build_session_created(self, conn_id: str) -> SessionCreatedEvent:
+        return self.session.build_session_created(conn_id)
+
+    def handle_session_update(self, conn_id: str, event: SessionUpdateEvent) -> Optional[RealtimeErrorEvent]:
+        return self.session.handle_session_update(conn_id, event)
+
     def handle_audio_append(self, conn_id: str, event: InputAudioBufferAppendEvent) -> list[bytes]:
-        """Decode base64 audio, resample to pipeline rate, and split into 512-sample PCM16 chunks for the VAD."""
-        try:
-            pcm_bytes = base64.b64decode(event.audio)
-        except Exception as e:
-            logger.error(f"Base64 decode error: {e}")
-            return []
-
-        audio_cfg = self.runtime_config.session.audio
-        if audio_cfg is not None and audio_cfg.input is not None:
-            client_in_rate = getattr(audio_cfg.input.format, "rate", None) or PIPELINE_SAMPLE_RATE
-        else:
-            client_in_rate = PIPELINE_SAMPLE_RATE
-        pcm_bytes = resample(pcm_bytes, client_in_rate, PIPELINE_SAMPLE_RATE)
-
-        st = self._state(conn_id)
-        pcm_bytes = st.audio_remainder + pcm_bytes
-
-        chunks = []
-        for i in range(0, len(pcm_bytes), CHUNK_SIZE_BYTES):
-            chunk = pcm_bytes[i : i + CHUNK_SIZE_BYTES]
-            if len(chunk) == CHUNK_SIZE_BYTES:
-                chunks.append(chunk)
-            else:
-                st.audio_remainder = chunk
-                break
-        else:
-            st.audio_remainder = b""
-
-        if chunks:
-            st.audio_buffer_has_data = True
-        return chunks
-
-    def handle_session_update(self, event: SessionUpdateEvent) -> Optional[RealtimeErrorEvent]:
-        """Apply session config changes to the shared RuntimeConfig.
-
-        Only ``RealtimeSessionCreateRequest`` sessions are accepted;
-        ``RealtimeTranscriptionSessionCreateRequest`` sessions not yet supported.
-        Incoming fields are deep-merged into the existing session so that
-        partial updates preserve previously-set values.
-        """
-        s = event.session
-        if s is None:
-            return None
-
-        if isinstance(s, RealtimeTranscriptionSessionCreateRequest):
-            return self.make_error(
-                message="Only 'realtime' session type is supported; transcription sessions are not.",
-                _type="invalid_session_type",
-            )
-
-        model = getattr(s, "model", None)
-        if model is not None:
-            logger.info(f"Session model set to: {model}")
-
-        current = self.runtime_config.session
-        if current is None:
-            self.runtime_config.session = s
-        else:
-            self.runtime_config.apply_session_update(s)
-        logger.info("Session configuration updated")
-        return None
-
-    def handle_conversation_item_create(self, conn_id: str, event: ConversationItemCreateEvent) -> list[ServerEvent]:
-        """Inject a text message or function-call output into the LLM context.
-
-        Items are added to the LLM chat context but do NOT trigger response
-        generation on their own.  A subsequent ``response.create`` event is
-        required to trigger the model.
-
-        Supported content part types:
-        - ``input_text``: added to LLM context (generation deferred to response.create).
-        - ``input_image``: forwarded to LLM context as a multi-part content list
-          alongside any ``input_text`` in the same message.
-        - ``function_call_output``: added to LLM context without triggering generation.
-        """
-        events: list[ServerEvent] = []
-        item = event.item
-
-        if not self._enqueue_item(item):
-            return []
-
-        if item:
-            st = self._state(conn_id)
-            events.append(
-                ConversationItemCreatedEvent(
-                    type="conversation.item.created",
-                    event_id=self._next_event_id(),
-                    previous_item_id=st.last_item_id,
-                    item=item,
-                )
-            )
-            st.last_item_id = item.id
-
-
-        return events
+        return self.audio.handle_audio_append(conn_id, event)
 
     def handle_audio_commit(self, conn_id: str) -> RealtimeErrorEvent | None:
-        """Commit the audio buffer. Returns an error if no audio was appended."""
-        st = self._state(conn_id)
-        if not st.audio_buffer_has_data:
-            return self.make_error(
-                message="Input audio buffer is empty, nothing to commit.",
-                _type="input_audio_buffer_commit_empty",
-            )
-        st.audio_buffer_has_data = False
-        logger.debug("Audio buffer committed")
-        return None
-
-    def handle_response_create(self, conn_id: str, event: ResponseCreateEvent) -> ServerEvent | None:
-        """Trigger a response.
-
-        Returns a ``ResponseCreatedEvent`` on success, a ``RealtimeErrorEvent``
-        on failure, or ``None`` if there is no text_prompt_queue.
-
-        The response is marked *in-progress* immediately so that a second
-        ``response.create`` arriving before the first audio chunk is correctly
-        rejected, and so that barge-in can cancel a pending response.
-
-        Per-response overrides (instructions, tool_choice) travel inside the
-        ``GENERATE_RESPONSE`` sentinel tuple so they are atomically paired
-        with the correct LLM generation — no shared mutable state to race on.
-
-        If ``event.response.input`` is present, its items are forwarded to the
-        LLM context before triggering generation (supports text + image items).
-        """
-        if self._state(conn_id).in_response:
-            return self.make_error(
-                message="Cannot create response while another response is in progress.",
-                _type="conversation_already_has_active_response",
-            )
-
-        override_instructions: str | None = None
-        override_tool_choice: str | None = None
-        if event.response:
-            if event.response.tool_choice:
-                if not isinstance(event.response.tool_choice, str):
-                    return self.make_error(
-                        message="Only string tool_choice values are supported for now (auto, required, none).",
-                        _type="tool_choice_not_supported",
-                    )
-                override_tool_choice = event.response.tool_choice
-                logger.info("Per-response tool_choice: %s", override_tool_choice)
-            cfg = self.runtime_config
-            if cfg.session is None:
-                cfg.session = RealtimeSessionCreateRequest(type="realtime")
-            if event.response.instructions:
-                override_instructions = event.response.instructions
-                logger.info(
-                    "Per-response instructions (%d chars): %s...",
-                    len(override_instructions),
-                    override_instructions[:60],
-                )
-            if hasattr(event.response, "metadata") and event.response.metadata:
-                self._state(conn_id).response_metadata = event.response.metadata
-                logger.info("Per-response metadata stored (%d keys)", len(event.response.metadata))
-            if event.response.input:
-                for input_item in event.response.input:
-                    self._enqueue_item(input_item)
-
-        st = self._state(conn_id)
-        st.in_response = True
-        st.current_response_id = _generate_id("resp")
-        self._start_item(conn_id)
-
-        if self.text_prompt_queue:
-            self.text_prompt_queue.put(
-                (MessageTag.GENERATE_RESPONSE, override_instructions, override_tool_choice)
-            )
-        logger.debug("response.create received, LLM generation triggered")
-        return ResponseCreatedEvent(
-            type="response.created",
-            event_id=self._next_event_id(),
-            response=self._build_response(conn_id, "in_progress"),
-        )
-
-    def _enqueue_item(self, item: ConversationItem) -> bool:
-        """Enqueue a single conversation item into the LLM context.
-
-        Returns ``True`` if the item was handled, ``False`` otherwise.
-        """
-        if not self.text_prompt_queue:
-            return False
-
-        if getattr(item, "type", None) == "message" and getattr(item, "content", None):
-            role = getattr(item, "role", None)
-            if not role or role not in ("user", "assistant"):
-                logger.warning("Unsupported message role: %s", role)
-                return False
-            content_parts: list[dict] = []
-            for part in item.content:
-                if (part.type == "input_text" and part.text) or (part.type == "input_image" and part.image_url):
-                    content_parts.append(part.model_dump(exclude_none=True))
-                else:
-                    logger.warning("Unsupported content part type: %s", part.type)
-            if content_parts:
-                self.text_prompt_queue.put((MessageTag.ADD_TO_CONTEXT, role, content_parts))
-                logger.debug("Enqueued message item (role=%s, %d parts)", role, len(content_parts))
-                return True
-            return False
-
-        if getattr(item, "type", None) == "function_call_output" and getattr(item, "output", None):
-            result_text = f"Call ID: {item.call_id}\nOutput: {item.output}"
-            self.text_prompt_queue.put((MessageTag.FUNCTION_RESULT, result_text))
-            logger.debug("Enqueued function_call_output (call_id=%s)", item.call_id)
-            return True
-
-        logger.warning("Unsupported item type: %s", getattr(item, "type", None))
-        return False
-
-    def handle_response_cancel(self, conn_id: str) -> list[ServerEvent]:
-        """Cancel the in-progress response and re-enable listening."""
-        events = self.finish_audio_response(conn_id, status="cancelled", reason="client_cancelled")
-        if self.should_listen:
-            self.should_listen.set()
-        logger.info("Response cancelled, listening re-enabled")
-        return events
-
-    # ── Outbound audio encoding ──────────────────
+        return self.audio.handle_audio_commit(conn_id)
 
     def encode_audio_chunk(self, conn_id: str, audio: bytes) -> list[ServerEvent]:
-        """Encode a raw PCM audio chunk, emitting ResponseCreated on the first chunk.
+        return self.audio.encode_audio_chunk(conn_id, audio)
 
-        When ``handle_response_create`` already allocated the response,
-        ``current_response_id`` is set and no duplicate event is emitted.
-        For the implicit-response path (VAD -> STT -> LLM -> TTS, no
-        ``response.create``), ``current_response_id`` is still ``None``
-        and the event is emitted here on the first audio chunk.
-        """
-        events: list[ServerEvent] = []
-        need_created = self._state(conn_id).current_response_id is None
-        resp_id, item_id = self._ensure_response(conn_id)
-        if need_created:
-            events.append(ResponseCreatedEvent(
-                type="response.created",
-                event_id=self._next_event_id(),
-                response=self._build_response(conn_id, "in_progress"),
-            ))
-        audio_cfg = self.runtime_config.session.audio
-        if audio_cfg is not None and audio_cfg.output is not None:
-            client_out_rate = getattr(audio_cfg.output.format, "rate", None) or PIPELINE_SAMPLE_RATE
-        else:
-            client_out_rate = PIPELINE_SAMPLE_RATE
-        audio = resample(audio, PIPELINE_SAMPLE_RATE, client_out_rate)
-        b64 = base64.b64encode(audio).decode("ascii")
-        events.append(ResponseAudioDeltaEvent(
-            type="response.output_audio.delta",
-            event_id=self._next_event_id(),
-            content_index=self._next_content_index(conn_id),
-            delta=b64,
-            item_id=item_id,
-            output_index=0,
-            response_id=resp_id,
-        ))
-        return events
+    def handle_response_create(self, conn_id: str, event: ResponseCreateEvent) -> ServerEvent | None:
+        return self.response.handle_response_create(conn_id, event)
+
+    def handle_response_cancel(self, conn_id: str) -> list[ServerEvent]:
+        return self.response.handle_response_cancel(conn_id)
 
     def finish_audio_response(
         self, conn_id: str, status: _ResponseStatus = "completed", reason: _StatusReason | None = None,
     ) -> list[ServerEvent]:
-        """Close the current response (audio done + response done)."""
-        st = self._state(conn_id)
-        events: list[ServerEvent] = []
-        if st.in_response:
-            resp_id, item_id = self._ensure_response(conn_id)
-            events.append(ResponseAudioDoneEvent(
-                type="response.output_audio.done",
-                event_id=self._next_event_id(),
-                content_index=0,
-                item_id=item_id,
-                output_index=0,
-                response_id=resp_id,
-            ))
-            events.append(ResponseDoneEvent(
-                type="response.done",
-                event_id=self._next_event_id(),
-                response=self._build_response(conn_id, status, reason),
-            ))
-            self._end_response(conn_id, status)
-        return events
+        return self.response.finish_audio_response(conn_id, status, reason)
 
-    # ── Pipeline text -> protocol events ─────────
+    def handle_conversation_item_create(self, conn_id: str, event: ConversationItemCreateEvent) -> list[ServerEvent]:
+        return self.conversation.handle_conversation_item_create(conn_id, event)
 
-    def translate_pipeline_text(self, conn_id: str, msg: dict) -> list[ServerEvent]:
-        """Convert an internal text_output_queue message to protocol ServerEvent(s)."""
-        events: list[ServerEvent] = []
+    def dispatch_pipeline_event(self, conn_id: str, msg: dict) -> list[ServerEvent]:
+        """Route a pipeline text_output_queue message to the appropriate handler."""
         msg_type = msg.get("type")
-
-        if msg_type == "speech_started":
-            if self._state(conn_id).in_response and self.runtime_config.interrupt_response_enabled:
-                events.extend(self.finish_audio_response(conn_id, status="cancelled", reason="turn_detected"))
-            input_item_id = self._start_item(conn_id)
-            st = self._state(conn_id)
-            st.last_item_id = input_item_id
-            st.response_usage.turns += 1
-            events.append(InputAudioBufferSpeechStartedEvent(
-                type="input_audio_buffer.speech_started",
-                event_id=self._next_event_id(),
-                audio_start_ms=msg.get("audio_start_ms", 0),
-                item_id=input_item_id,
-            ))
-
-        elif msg_type == "speech_stopped":
-            duration_s = msg.get("duration_s", 0.0)
-            if duration_s:
-                self._state(conn_id).input_audio_duration_s = duration_s
-            events.append(InputAudioBufferSpeechStoppedEvent(
-                type="input_audio_buffer.speech_stopped",
-                event_id=self._next_event_id(),
-                audio_end_ms=msg.get("audio_end_ms", msg.get("audio_stop_ms", 0)),
-                item_id=self._current_item_id(conn_id),
-            ))
-
-        elif msg_type == "assistant_text":
-            resp_id, item_id = self._ensure_response(conn_id)
-            self._state(conn_id).last_item_id = item_id
-            output_idx = 0
-            text = msg.get("text", "")
-            if text:
-                events.append(ResponseAudioTranscriptDoneEvent(
-                    type="response.output_audio_transcript.done",
-                    event_id=self._next_event_id(),
-                    content_index=0,
-                    item_id=item_id,
-                    output_index=output_idx,
-                    response_id=resp_id,
-                    transcript=text,
-                ))
-                output_idx += 1
-            tools = msg.get("tools")
-            if tools:
-                self._state(conn_id).response_usage.tool_calls += len(tools)
-                for tool in tools:
-                    if isinstance(tool.get("arguments"), str):
-                        arguments = tool.get("arguments")
-                    else:
-                        arguments = json.dumps(tool.get("arguments", {}))
-                    events.append(
-                        ResponseFunctionCallArgumentsDoneEvent(
-                            type="response.function_call_arguments.done",
-                            event_id=self._next_event_id(),
-                            call_id=tool.get("call_id", ""),
-                            name=tool.get("name", ""),
-                            arguments=arguments,
-                            item_id=item_id,
-                            output_index=output_idx,
-                            response_id=resp_id,
-                        )
-                    )
-                    output_idx += 1
-
-        elif msg_type == "partial_transcription":
-            events.append(
-                ConversationItemInputAudioTranscriptionDeltaEvent(
-                    type="conversation.item.input_audio_transcription.delta",
-                    event_id=self._next_event_id(),
-                    content_index=self._next_content_index(conn_id),
-                    item_id=self._current_item_id(conn_id),
-                    delta=msg.get("delta", ""),
-                )
-            )
-
-        elif msg_type == "transcription_completed":
-            st = self._state(conn_id)
-            st.response_usage.audio_duration_s += st.input_audio_duration_s
-            events.append(
-                ConversationItemInputAudioTranscriptionCompletedEvent(
-                    type="conversation.item.input_audio_transcription.completed",
-                    event_id=self._next_event_id(),
-                    content_index=0,
-                    item_id=self._current_item_id(conn_id),
-                    transcript=msg.get("transcript", ""),
-                    usage=UsageTranscriptTextUsageDuration(
-                        seconds=st.input_audio_duration_s, type="duration",
-                    ),
-                )
-            )
-
-        elif msg_type == "token_usage":
-            st = self._state(conn_id)
-            st.response_usage.input_tokens += msg.get("input_tokens", 0)
-            st.response_usage.output_tokens += msg.get("output_tokens", 0)
-            logger.info(
-                "Token usage (response): input=%d, output=%d",
-                st.response_usage.input_tokens, st.response_usage.output_tokens,
-            )
-
-        else:
-            logger.debug(f"Unhandled pipeline text message type: {msg_type}")
-
-        return events
+        handler = self._pipeline_dispatch.get(msg_type)
+        if handler is None:
+            logger.debug("Unhandled pipeline message type: %s", msg_type)
+            return []
+        return handler(conn_id, msg)
 
     # ── Metrics ────────────────────────────────────
+
+    def _on_token_usage(self, conn_id: str, msg: dict) -> list[ServerEvent]:
+        """Accumulate input/output token counts on the connection's usage metrics."""
+        st = self._state(conn_id)
+        st.response_usage.input_tokens += msg.get("input_tokens", 0)
+        st.response_usage.output_tokens += msg.get("output_tokens", 0)
+        logger.info(
+            "Token usage (response): input=%d, output=%d",
+            st.response_usage.input_tokens, st.response_usage.output_tokens,
+        )
+        return []
 
     def get_usage(self) -> dict:
         """Return cumulative usage metrics across all completed responses."""
@@ -708,7 +279,7 @@ class RealtimeService:
         data["total_errors"] = self.total_usage.total_errors
         return data
 
-    # ── Helpers ───────────────────────────────────
+    # ── Error ───────────────────────────────────
 
     def make_error(self, message: str, _type: str) -> RealtimeErrorEvent:
         self.total_usage.record_error(_type)
